requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
Pillow==8.3.2
rembg==1.0.28